requirements.txt
python-dotenv
python-telegram-bot
sqlalchemy
google-generativeai
streamlit
groq
pyTelegramBotAPI
